If apply_all_transactions is missing data, reraise
Just continuing is all kinds of wrong. I don't know how this bug lasted
so long. I guess it's not tested or used often.

It caused some unfortunate bugs for @lithp that were hard to track down,
because the state got reverted, but the receipt was still added. Sorry,
buddy!